rename src/ to project name
